refactor(craft): move the craft MCP hook into the tools feature lib
`web/src/hooks/` is for general-purpose hooks; a hook dedicated to the craft MCP
endpoint belongs with the rest of the MCP code, next to `mcpService` and
`interfaces`. Matches the `lib/<feature>/hooks.ts` convention used by agents,
auth, analytics and others, and switches to a named export to match.
